Use the new MiniTest::Spec::DSL.
Removed all Test::Unit monkey patches in favor of MiniTestSpecRails::DSL. Our DSL overrides MiniTest's `before` and `after` and delegates to ActiveSupport's SetupAndTeardown callbacks. This was necessary to keep a single responsibility for setup and teardown, otherwise you could not guarantee which would be run first. If people only used the MiniTest::Spec DSL, this would not be an issue. But since this gem aims to be a drop in replacement for legacy applications, we do it and we only do it in scope of ActiveSupport's usage of the DSL. Lastly, ActiveSupport's callbacks are stackable, so you can setup multiple times using a symbol to that method.

* Instllled guard-minitest and minitest-focus for dev help.