Ensure spec methods are created with correct source location
Previously, examining the source location of a spec method created using
the `test` keyword provided by MiniTestSpecRails::DSL would show the
method as being defined in the minitest-spec-rails gem. Ideally we would
want this to show the actual location where the test is defined, not a
gem method. Other gems depend on the source location yielding correct
information.

We can solve this by reordering the `instance_eval` and `it` method
calls inside the `test` DSL method.